fix layout
- fix all-news card scroll
- remove unnecessary dom nesting
