+++ apps/frontend/src/routes/{-$lang}/_layout.tsx
@@ -9,6 +9,7 @@ import { IntlProvider } from "use-intl";
 import { i18n, localeSchema } from "@/config/i18n";
 import { $getMessages } from "@/serverFunctions/i18n";
 import { $getSiteNavigation } from "@/serverFunctions/siteNavigation";
+import { Navbar } from "@/components/Navbar";
 
 // import { getMessagesFn } from "@/serverFunctions/locale";
 
@@ -71,6 +72,7 @@ function RouteComponent() {
       timeZone={timeZone}
     >
       <CatchBoundary getResetKey={() => "reset"}>
+        <Navbar />
         <Outlet />
       </CatchBoundary>
     </IntlProvider>
