+++ apps/frontend/src/routes/{-$lang}/_layout/_main.tsx
@@ -30,10 +30,8 @@ export const Route = createFileRoute("/{-$lang}/_layout/_main")({
 function RouteComponent() {
   return (
     <>
-      <main className="z-10 flex flex-1 flex-col gap-4 p-4">
-        <Navbar />
+      <main className="z-10 flex flex-1 flex-col items-stretch gap-4 p-4">
         <Alerts />
-
         <Outlet />
       </main>
       <Footer />
